requirements.txt
fastapi
uvicorn
openai
python-dotenv
sqlalchemy
requests
beautifulsoup4
pydantic
